refactor: JpaConfig 공통 모듈로 이동
