+++ common/src/main/java/com/flight_booking/common/infrastructure/auditing/JpaConfig.java
@@ -1,0 +1,17 @@
+package com.flight_booking.common.infrastructure.auditing;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+
+  @Bean
+  public AuditorAware<String> auditorProvider() {
+    return new AuditorAwareImpl();
+  }
+
+}
