--- flight-service/src/main/java/com/flight_booking/flight_service/infrastructure/configuration/JpaConfig.java
@@ -1,25 +1,0 @@
-package com.flight_booking.flight_service.infrastructure.configuration;
-
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.Optional;
-import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaConfig {
-
-  @Bean
-  public AuditorAware<String> auditorAware() {
-    return () -> {
-      //TODO : 추후 유저 구현되면 수정
-      String userId = "tmpUser";
-
-      return Optional.of(userId);
-    };
-  }
-}
