--- payment-service/src/main/java/com/flight_booking/payment_service/infrastructure/configuration/JpaConfig.java
@@ -1,32 +1,0 @@
-package com.flight_booking.payment_service.infrastructure.configuration;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import java.util.Optional;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-
-  @PersistenceContext
-  private EntityManager entityManager;
-
-  @Bean
-  public JPAQueryFactory jpaQueryFactory() {
-    return new JPAQueryFactory(entityManager);
-  }
-
-  @Bean
-  public AuditorAware<String> auditorAware() {
-    return () -> {
-      String email = "temp"; // TODO
-
-      return Optional.of(email);
-    };
-  }
-}
