+++ user-service/src/main/java/com/flight_booking/user_service/infrastructure/auditing/JpaConfig.java
@@ -1,6 +1,5 @@
-package com.flight_booking.ticket_service.infrastructure.configuration;
+package com.flight_booking.user_service.infrastructure.auditing;
 
-import java.util.Optional;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
@@ -11,12 +10,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 public class JpaConfig {
 
   @Bean
-  public AuditorAware<String> auditorAware() {
-    return () -> {
-      //TODO : 추후 유저 구현되면 수정
-      String userId = "tmpUser";
-
-      return Optional.of(userId);
-    };
+  public AuditorAware<String> auditorProvider() {
+    return new AuditorAwareImpl();
   }
 }
