--- user-service/src/main/java/com/flight_booking/user_service/infrastructure/persistence/JpaConfig.java
@@ -1,10 +1,0 @@
-package com.flight_booking.user_service.infrastructure.persistence;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaConfig {
-
-}
